requirements.txt
aiofiles==23.1.0
aiosmtplib==2.0.2
amqp==5.1.1
anyio==3.7.0
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-mail==1.3.0
flower==2.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
humanize==4.6.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
mypy==1.4.1
mypy-extensions==1.0.0
passlib==1.7.4
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
